requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
bcrypt
itsdangerous
python-multipart
pyotp
qrcode[pil]
webauthn

# Hardening
slowapi
